Fix replacement paths that are absolute paths on Windows
There is a bug where absolute paths have their backslashes replaced as forward slashes.  For relative paths, this is not a problem.  However, especially with joeybaker/remapify, it is possible to have absolute paths passed into aliasify, and getting the backslashes replaced there is problematic.

This fixes the issue by testing for an absolute path on Windows, and then replacing backslashes with double backslashes, which is necessary since the filename is in quotes.  All other paths have their backslashes replaced by forward slashes as currently happens today.